Move test fixtures to new herostratus-tests crate
This enables sharing them between herostratus proper and the integration
tests, while keeping them out of the public API of the herostratus
library.
