+++ herostratus-tests/Cargo.toml
@@ -6,7 +6,7 @@ license.workspace = true
 version.workspace = true
 edition.workspace = true
 
-[dev-dependencies]
+[dependencies]
 herostratus.path = "../herostratus"
 
 assert_cmd.workspace = true
@@ -14,7 +14,9 @@ ctor.workspace = true
 eyre.workspace = true
 git2.workspace = true
 lazy_static.workspace = true
-predicates.workspace = true
 tempfile.workspace = true
 tracing-subscriber.workspace = true
 tracing.workspace = true
+
+[dev-dependencies]
+predicates.workspace = true
