+++ herostratus-tests/src/fixtures/mod.rs
@@ -1,3 +1,2 @@
 pub mod config;
 pub mod repository;
-pub mod rule;
