+++ herostratus-tests/tests/check_all.rs
@@ -1,19 +1,17 @@
-mod common;
-
-use common::CommandExt;
+use herostratus_tests::cmd::{herostratus, CommandExt};
 use predicates::prelude::*;
 use predicates::str;
 
 #[test]
 fn add_self_and_then_check_all() {
     let self_dir = format!("file://{}/..", env!("CARGO_MANIFEST_DIR"));
-    let (mut cmd, temp) = common::herostratus(None);
+    let (mut cmd, temp) = herostratus(None);
     cmd.arg("add").arg("--skip-clone").arg(self_dir);
 
     let output = cmd.captured_output().unwrap();
     assert!(output.status.success());
 
-    let (mut cmd, _) = common::herostratus(Some(temp.as_ref().unwrap().path()));
+    let (mut cmd, _) = herostratus(Some(temp.as_ref().unwrap().path()));
     // If 'add' skips the clone, using 'fetch-all' or 'check-all' without '--no-fetch' will clone
     cmd.arg("check-all");
     let output = cmd.captured_output().unwrap();
