+++ herostratus/src/lib.rs
@@ -4,5 +4,3 @@ pub mod commands;
 pub mod config;
 pub mod git;
 pub mod rules;
-#[cfg(test)]
-pub mod test;
