+++ herostratus/src/rules/h002_shortest_subject_line.rs
@@ -78,9 +78,10 @@ impl Rule for ShortestSubjectLine {
 
 #[cfg(test)]
 mod tests {
+    use herostratus_tests::fixtures;
+
     use super::*;
     use crate::achievement::grant_with_rules;
-    use crate::test::fixtures;
 
     #[test]
     fn test_all_above_threshold() {
